+++ src/Domain/Module/Forgejo.hs
@@ -9,6 +9,7 @@ import Data.Text (Text)
 import qualified Data.Text as T
 import Domain.Error (renderDomainError)
 import Domain.Module (ModuleName, mkModuleName, moduleNameText)
+import Policy (Policy)
 
 data ForgejoConfig = ForgejoConfig
     { forgejoConfigName :: ModuleName
@@ -16,6 +17,7 @@ data ForgejoConfig = ForgejoConfig
     , forgejoHttpPort :: Maybe Int
     , forgejoSshPort :: Maybe Int
     , forgejoDomain :: Maybe Text
+    , forgejoPolicies :: [Policy]
     }
     deriving (Eq, Show)
 
@@ -26,6 +28,7 @@ instance ToJSON ForgejoConfig where
                 <> maybe [] (\p -> ["http_port" .= p]) (forgejoHttpPort fc)
                 <> maybe [] (\p -> ["ssh_port" .= p]) (forgejoSshPort fc)
                 <> maybe [] (\d -> ["domain" .= d]) (forgejoDomain fc)
+                <> (if null (forgejoPolicies fc) then [] else ["policies" .= forgejoPolicies fc])
 
 instance FromJSON ForgejoConfig where
     parseJSON = withObject "ForgejoConfig" $ \o -> do
@@ -36,3 +39,4 @@ instance FromJSON ForgejoConfig where
             <*> o .:? "http_port"
             <*> o .:? "ssh_port"
             <*> o .:? "domain"
+            <*> o .:? "policies" .!= []