+++ src/Domain/Module/PostgreSQL.hs
@@ -9,6 +9,7 @@ import Data.Text (Text)
 import qualified Data.Text as T
 import Domain.Error (renderDomainError)
 import Domain.Module (ModuleName, mkModuleName, moduleNameText)
+import Policy (Policy)
 
 data PostgreSQLConfig = PostgreSQLConfig
     { postgresqlConfigName :: ModuleName
@@ -16,6 +17,7 @@ data PostgreSQLConfig = PostgreSQLConfig
     , postgresqlPort :: Maybe Int
     , postgresqlDataDir :: Maybe Text
     , postgresqlMaxConnections :: Maybe Int
+    , postgresqlPolicies :: [Policy]
     }
     deriving (Eq, Show)
 
@@ -26,6 +28,7 @@ instance ToJSON PostgreSQLConfig where
                 <> maybe [] (\p -> ["port" .= p]) (postgresqlPort pc)
                 <> maybe [] (\d -> ["data_dir" .= d]) (postgresqlDataDir pc)
                 <> maybe [] (\m -> ["max_connections" .= m]) (postgresqlMaxConnections pc)
+                <> (if null (postgresqlPolicies pc) then [] else ["policies" .= postgresqlPolicies pc])
 
 instance FromJSON PostgreSQLConfig where
     parseJSON = withObject "PostgreSQLConfig" $ \o -> do
@@ -36,3 +39,4 @@ instance FromJSON PostgreSQLConfig where
             <*> o .:? "port"
             <*> o .:? "data_dir"
             <*> o .:? "max_connections"
+            <*> o .:? "policies" .!= []